requirements.txt
python-binance
typer[all]
rich
loguru
python-dotenv
